+++ CHANGELOG.md
@@ -75,6 +75,8 @@ Adding a new version? You'll need three changes:
 - Telemetry reports now include a number of discovered Gateways when the Gateway Discovery
   feature is turned on.
   [#3783](https://github.com/Kong/kubernetes-ingress-controller/pull/3783)
+- `HTTPRoute` reconciler now watches relevant `ReferenceGrant`s for changes.
+  [#3759](https://github.com/Kong/kubernetes-ingress-controller/pull/3759)
 
 ### Fixed